Add files via upload
Adding PNGs used in the footer for each page. NOTE: When viewing the website, the PNGs might have to be in the same folder as the html files in order to see the images (if not, the path to the images will probably have to be updated in the each individual html file).